backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
slowapi==0.1.9
alembic==1.13.1
python-dotenv==1.0.0
pytz==2023.3
fastapi-cache2[redis]==0.2.1
redis==4.6.0
apscheduler==3.10.4
python-magic==0.4.27
Pillow==10.2.0
openpyxl==3.1.2
pandas==2.1.4
prometheus-fastapi-instrumentator==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0
faker==22.0.0
ImageHash==4.3.1
psycopg2-binary==2.9.9
